+++ apps/meteor/app/api/server/default/openApi.ts
@@ -1,3 +1,4 @@
+import type { Route } from '@rocket.chat/http-router';
 import { isOpenAPIJSONEndpoint } from '@rocket.chat/rest-typings';
 import express from 'express';
 import { WebApp } from 'meteor/webapp';
@@ -6,7 +7,6 @@ import swaggerUi from 'swagger-ui-express';
 import { settings } from '../../../settings/server';
 import { Info } from '../../../utils/rocketchat.info';
 import { API } from '../api';
-import type { Route } from '../router';
 
 const app = express();