+++ apps/meteor/client/lib/rooms/roomTypes/public.ts
@@ -1,7 +1,7 @@
 import type { AtLeast, IRoom } from '@rocket.chat/core-typings';
 import { isRoomFederated } from '@rocket.chat/core-typings';
 import { Meteor } from 'meteor/meteor';
-import type { Mongo } from 'meteor/mongo';
+import type { Filter } from 'mongodb';
 
 import { hasAtLeastOnePermission } from '../../../../app/authorization/client';
 import { Rooms } from '../../../../app/models/client';
@@ -109,7 +109,7 @@ roomCoordinator.add(
 		},
 
 		findRoom(identifier) {
-			const query: Mongo.Selector<IRoom> = {
+			const query: Filter<IRoom> = {
 				t: 'c',
 				name: identifier,
 			};